+++ package.json
@@ -46,13 +46,11 @@
 	"dependencies": {
 		"@floating-ui/core": "^1.5.0",
 		"@melt-ui/svelte": "^0.61.1",
-		"@melt-ui/pp": "^0.1.4",
 		"esm-env": "^1.0.0",
 		"just-debounce-it": "^3.2.0",
 		"just-throttle": "^4.2.0",
 		"lucide-svelte": "^0.292.0",
 		"svelte-headlessui": "^0.0.32",
-		"svelte-sequential-preprocessor": "^2.0.1",
 		"tailwind-merge": "^2.0.0"
 	},
 	"peerDependencies": {
