+++ src/lib/collapsible/Collapsible.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-	import { actions, type UseActions } from '$lib/actions';
-	import { createCollapsible, melt } from '@melt-ui/svelte';
+	import { createCollapsible } from '@melt-ui/svelte';
 	import { slide } from 'svelte/transition';
+	import { actions, type UseActions } from '$lib/actions';
 	import { ChevronDown } from 'lucide-svelte';
 	import { twMerge } from 'tailwind-merge';
 
@@ -17,8 +17,12 @@
 </script>
 
 <div class={twMerge('bg-white shadow rounded-xl', className)} use:actions={use} {...$$restProps}>
-	<div use:melt={$root}>
-		<div class="flex hover:cursor-pointer hover:bg-gray-50 rounded-xl" use:melt={$trigger}>
+	<div {...$root} use:root>
+		<div
+			class="flex hover:cursor-pointer hover:bg-gray-50 rounded-xl justify-between"
+			{...$trigger}
+			use:trigger
+		>
 			<slot name="label" />
 			<div
 				class={twMerge('place-self-end sm:place-self-center max-sm:mb-2 mx-3', iconClass)}
@@ -30,7 +34,7 @@
 
 		<div style:top="calc(100% + 10px)" style:right="0" style:left="0">
 			{#if $open}
-				<div use:melt={$content} transition:slide>
+				<div {...$content} use:content transition:slide>
 					<slot />
 				</div>
 			{/if}
