+++ svelte.config.js
@@ -1,7 +1,5 @@
 import adapter from '@sveltejs/adapter-cloudflare';
 import { vitePreprocess } from '@sveltejs/kit/vite';
-import { preprocessMeltUI } from '@melt-ui/pp';
-import sequence from 'svelte-sequential-preprocessor';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
@@ -11,10 +9,7 @@ const config = {
 	},
 	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
 	// for more information about preprocessors
-	preprocess: sequence([
-		vitePreprocess(),
-		preprocessMeltUI(), // keep this in the last order
-	]),
+	preprocess: vitePreprocess(),
 
 	kit: {
 		csp: {
